Use the streaming HttpEngine by default on Apple
