--- tests/defer/src/appleTest/kotlin/test/getStreamingHttpEngine.kt
@@ -1,6 +1,0 @@
-package test
-
-import com.apollographql.apollo3.network.http.HttpEngine
-import com.apollographql.apollo3.network.http.StreamingNSURLSessionHttpEngine
-
-actual fun getStreamingHttpEngine(): HttpEngine = StreamingNSURLSessionHttpEngine()
