--- tests/defer/src/jsTest/kotlin/test/getStreamingHttpEngine.kt
@@ -1,6 +1,0 @@
-package test
-
-import com.apollographql.apollo3.network.http.DefaultHttpEngine
-import com.apollographql.apollo3.network.http.HttpEngine
-
-actual fun getStreamingHttpEngine(): HttpEngine = DefaultHttpEngine()
