--- tests/multipart/src/appleTest/kotlin/test/MultipartSubscriptionsMockServerTest.apple.kt
@@ -1,8 +1,0 @@
-package test
-
-import com.apollographql.apollo3.network.http.HttpEngine
-import com.apollographql.apollo3.network.http.StreamingNSURLSessionHttpEngine
-
-actual fun StreamingHttpEngine(): HttpEngine {
-  return StreamingNSURLSessionHttpEngine()
-}
