--- tests/multipart/src/jsTest/kotlin/test/MultipartSubscriptionsMockServerTest.js.kt
@@ -1,7 +1,0 @@
-package test
-
-import com.apollographql.apollo3.network.http.HttpEngine
-
-actual fun StreamingHttpEngine(): HttpEngine {
-  TODO("Not yet implemented")
-}
