--- tests/multipart/src/jvmTest/kotlin/test/MultipartSubscriptionsMockServerTest.jvm.kt
@@ -1,8 +1,0 @@
-package test
-
-import com.apollographql.apollo3.network.http.DefaultHttpEngine
-import com.apollographql.apollo3.network.http.HttpEngine
-
-actual fun StreamingHttpEngine(): HttpEngine {
-  return DefaultHttpEngine()
-}
